requirements.txt
asgiref==3.4.1
awesome-slugify==1.6.5
Django==3.2.9
django-ckeditor==6.1.0
django-extensions==3.1.3
django-js-asset==1.2.2
djangorestframework==3.12.4
graphviz==0.17
Pillow==8.4.0
pydotplus==2.0.2
pyparsing==3.0.4
pytz==2021.3
regex==2021.11.10
sqlparse==0.4.2
Unidecode==0.4.21
